Made out_port purely combinatorial.
PicoBlaze reads from its in_port on the same cycle that port_id is asserted, (the cycle *before* read_strobe).
Since out_port was being assigned in a process, it was a cycle late.  Making it combinatorial corrects this.

We can now read and write registers in the multiplier *on the FPGA!*